Show one decimal place for percentage changes
- getChangePercent() now returns double instead of int
- Format strings updated to %.1f%% in overlay and panel
- Sorting comparator updated to comparingDouble

+++ src/main/java/com/gepriceoverlay/PriceData.java
@@ -17,10 +17,10 @@ public class PriceData
 
 	public int getChange() { return current - previous; }
 
-	public int getChangePercent()
+	public double getChangePercent()
 	{
 		if (previous == 0) return 0;
-		return (int) ((current - previous) / (double) previous * 100);
+		return (current - previous) / (double) previous * 100;
 	}
 
 	public static String formatGold(int change)
